==> 05-miyasaka2014/01-download-zips.R <==
## This script assumed that you have run the file "05-miyasaka2014/00-setup.R"


urls <- "http://flybrain.mrc-lmb.cam.ac.uk/si/nat/miyasaka/pruned-traces.zip"

message("Downloading data (1.6 MB) if necessary...")
for (url in urls){
  localfile <- basename(url)
  if(file.exists(localfile)) next
  message("Downloading ", url, "...")
  t <- try(download.file(url, localfile, mode='wb'))
  if(inherits(t, 'try-error')) {
    message("Unable to download ", url)
    next
  }
}

message("Unzipping")
unzip("pruned-traces.zip")
